requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1
openai>=1.0.0
qdrant-client>=1.9.0
fastembed>=0.3.0
pypdf>=4.0.0
supabase>=2.4.0
langsmith>=0.1.0
reportlab>=4.0.0
